Fix OpenTelemetry span processing when using logstasher
The logstasher gem monkey patches the `process_action.action_controller`
instrumentation. However the gem hasn't been updated for Rails 7, which
assumes that the payload has the 'request' object available. This
request object is used by the OpenTelemetry SDK to process spans. This
generates lots of error logs from the SDK, and prevents us from
recording some spans.

This re-monkey patches the instrumentation to use a raw_payload that is
expected in Rails 7.  Ideally, this would be fixed in the logstasher gem
however there has been little active development over the last year.
Longer term fix might be to move to a maintained logging library.