Set time_zone to London in all GOV.UK apps
Our conventions for rails apps say that we should set time_zone to
London:

https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#specify-london-as-the-timezone

In practice however, only about half of the apps do so. Based on the
apps running in integration:

Apps which set the time_zone to London:

    AccountApi
    Collections
    CollectionsPublisher
    ContentPublisher
    Collections
    Frontend
    GovernmentFrontend
    Static
    EmailAlertApi
    Frontend
    GovernmentFrontend
    GovukChat
    LocationsApi
    Maslow
    PlacesManager
    Publisher
    ReleaseApp
    SearchApiV2
    Signon
    SmartAnswers
    SpecialistPublisher
    Static
    Whitehall

Apps which leave the time_zone as it's default UTC:

    AuthenticatingProxy
    Contacts
    ContentDataAdmin
    ContentPerformanceManager
    ContentStore
    ContentTagger
    ContentStore
    RouterApi
    EmailAlertFrontend
    Feedback
    FinderFrontend
    HmrcManualsApi
    LinkCheckerApi
    LocalLinksManager
    ManualsPublisher
    RouterApi
    SearchAdmin
    SearchV2Evaluator
    ServiceManualPublisher
    ShortUrlManager
    Support
    SupportApi
    Transition
    TravelAdvicePublisher

Having the apps use inconsistent time_zones feels like a recipe for
trouble.

config.time_zone has to be set before
active_support.initialize_time_zone runs, so we specify that our
initializer must run before it. Unfortunately this means that there's no
way for applications to set their own value of time_zone. I think this
is probably desirable for consumers of govuk_app_config though - I can't
think of a legitimate reason why we'd want some apps using different
time_zones.

Technically this is a breaking change, as it could change behaviour in
apps which are using timezone-specific methods and expecting to get UTC
times out. Hopefully we're already explicit about when we want UTC and
when we want local time though.

Once this is merged, I'll remove the bit in the docs about setting
time_zone, as it will be automatic.