Cut release for London Time zone
I've made this a minor version, even though it's technically a breaking change (or at least, updating will result in a change in behaviour).

I'm pretty confident that the new behaviour (London time_zone instead of UTC) is correct for all GOV.UK apps, so they shouldn't need to think about this as a "breaking" change.