Configure time_zone with govuk_time_zone
In the original implementation of the GovukTimezone module we assumed
that literally all GOV.UK apps would use the London timezone. We didn't
really consider how to make it easy for apps to override this if needed.

Since then, we've found a couple of examples where apps genuinely do
want to override the timezone:

- publishing-api, which always wants to present dates as UTC because its
  an API
- travel-advice-publisher, which always wants to present dates as UTC
  because a majority of its users are in non-London timezones (so BST is
  confusing)

Currently, these apps override GovukTimezone by inserting an initializer
between it and active_support.initialize_time_zone like this:

    initializer "publishing_api.configure_timezone",
                after: "govuk_app_config.configure_timezone",
                before: "active_support.initialize_time_zone" do
      config.time_zone = "UTC"
    end

    # https://github.com/alphagov/publishing-api/blob/252f47a1/config/application.rb#L123-L129

This is a bit of a faff, and it also results in some very confusing
logs.

Because the GovukTimezone still runs in this setup, it would log
"govuk_app_config changing time_zone from UTC (the default) to London"
... but then we'd set it back to UTC without logging anything, making it
seem from the logs that the time_zone was London when it was actually
UTC. Bleurgh.

Unfortunately, it's not possible to work out just from config.time_zone
whether the time_zone is set to UTC because its the default (in which
case we should override it to London), or because the app has explicitly
set it (in which case we should leave it alone).

I think the best approach is to introduce a new method govuk_time_zone,
which won't be set by default to allow us to distinguish these
scenarios.

This has the nice benefit that we can be a bit more strict about what's
an error case and what just requires a log message.

If config.time_zone is set to anything other than UTC, that's an error.
If config.govuk_time_zone is set, we use that, otherwise we override UTC
to London and log a message.

This will be slightly annoying to release, because we'll have to remove
all the instances of config.time_zone = "London" across our apps. If we
want to be explicit, we can set config.govuk_time_zone. If we're happy
with the default, then we can just remove the line.