Update SidekiqRedis to work with Sidekiq 7
This healthcheck has been failing for applications that have upgraded to Sidekiq 7.
Sidekiq 7 uses 'redis-client' which has a different API to 'redis'
which was used in Sidekiq 6.

To access redis_info you now have to call

```
Sidekiq.default_configration.redis_info
```

while in Sidekiq 6 you call

```
Sidekiq.redis_info
```

This updates the healthcheck to work with both versions of Sidekiq checking
if Sidekiq.responds to :redis_info.

If it does it calls that, otherwise it calls Sidekiq.default_configration.redis_info.

It's not missively clear these healthchecks are actually being used
for anything since this has been failing on production applications for
a couple of weeks without anyone noticing.

We're in the process of speaking to the platform team to see if they're
actually being used for anything and if not whether they should be removed.

But for now i'm just updating it to work with Sidekiq 7.