Release v9.18.0: adds 'GovukEnvironment.current' feature
As discussed in Tech Fortnightly on 22nd July:
https://docs.google.com/document/d/1X-AberK3K4XzXxY6N4zzKZJY5t-bXfi5YX4SAcLpYtk/edit?disco=AAABnYvu2y8

We lacked a 'canonical' way for an app to determine what environment
of GOV.UK it is running on.

- This used to be achieved by a call to the (now deprecated)
  [GovukAdminTemplate.environment_label](https://github.com/alphagov/govuk_admin_template/blob/457111d71cd2ee530ab69076ce53f623b59825bd/lib/govuk_admin_template.rb#L14-L16)
  (and in any case, was only available to publishing apps; the
  consumers of said 'admin templates')
- There is also a [SENTRY_CURRENT_ENV](https://docs.publishing.service.gov.uk/repos/govuk_app_config.html#automatic-error-reporting)
  definition which does the job, but it being tied to the name of
  a third party service doesn't inspire confidence in its
  longevity.
- There's the [GOVUK_ENVIRONMENT](https://github.com/alphagov/govuk-helm-charts/blob/6a3432cf2e1ac0fa27b65feffefd7a8b72caf224/charts/generic-govuk-app/values.yaml#L148)
  env variable, which seems to be exactly what we want: but there
  was general agreement at Tech Fortnightly that we should have
  a proper 'common library' way of querying the environment, not
  relying on an environment variable that could be renamed (or
  overwritten at runtime).
- There's also reference to [GOVUK_ENVIRONMENT_NAME](https://github.com/alphagov/datagovuk_find/blob/a9f6bf29b96ce50e2b260b2289068bedf9852ed0/config/initializers/sentry.rb#L6)
- There's also the [current_acceptance_environment method](https://github.com/alphagov/govuk_publishing_components/blob/86582986ea6024a4a199300084b2a65ae48a3767/lib/govuk_publishing_components/app_helpers/environment.rb#L8C16-L8C46)
  in govuk_publishing_components, but that is curiously named and,
  again, doesn't universally apply to all apps (not every app
  pulls in that gem), whereas govuk_app_config _does_ apply to
  every app - so is a much better home for that sort of querying.
- And, until recently, some apps also queried the GOVUK_WEBSITE_ROOT
  env var or looked at the current URL. See https://github.com/alphagov/finder-frontend/commit/0b121308709b19cb9849c0fffbc8a92b29a063ee

All this to say, there are many ways one can determine one's
environment, and we should [settle on one true way](https://xkcd.com/927/).

The discussion did also look more broadly at whether or not
querying one's current environment is a good idea in the first
place. Code that does so may only break on production, for example,
because it has a broken branch of logic that is only actioned when
running in that environment. There's a school of thought that
everything should be represented as toggleable 'features' instead,
that can be independently tested irrespective of their environment.

I would argue that that is true, but that there are some use cases
where knowing your environment is important (think Release app,
EKS integrations etc), and that settling on one way of querying the
environment is less smelly than letting all apps take their own
haphazard approaches (duplication of effort, etc).